Add spotifyRedesign and Demos
