--- Application/README.md
@@ -1,7 +1,0 @@
-# Spotify Redesign - React Native mobile application
-
-React Native app made and tested using Expo Snack. The app contains a dynamic and interactive implementation of the figma UI mockup (find in ./Redesign Approach/wireframe.svg), while also combining the features shown in the user flow diagram. (find in ./Redesign Approach/Enhanced-InputSupport User Flow Diagram.pdf)
-
-You may also view and simulate the Snack on multiple devices here: [Expo](https://snack.expo.dev/@victorgooge/spotifyredesign)
-
-Snack is Open Source. You can find the code on the [GitHub repo](https://github.com/expo/snack).
